refactor: features/userManagement 제거
