+++ frontend/techpick/src/utils/index.ts
@@ -1,2 +1,3 @@
 export { notifyError, notifySuccess } from './toast';
 export { numberToRandomColor } from './color';
+export { getClientCookie } from './getClientCookie';
